Fix null ref in L0 (#4274)

+++ Tests-Legacy/L0/lib/_suite.ts
@@ -22,7 +22,9 @@ describe('Test Helpers Suite', function() {
     });
 
     after(function () {
-        psr.kill();
+        if (psr) {
+            psr.kill();
+        }
     });
 
     if (psm.testSupported()) {
